Avoid calculating self.etc_chef_dir multiple times on startup
Running chef-apply we calculated this value 5 times. We also have several other resources that use this method which would again have to be calculated. Each of those calculations run the PathHelper.join which involves a whole series of terrible string manipulations. Just calculate this once and be done with it.

Signed-off-by: Tim Smith <tsmith@chef.io>